Split up pratt/types
