+++ frontend/src/metabase/querying/expressions/pratt/index.ts
@@ -2,4 +2,5 @@ export * from "./lexifier";
 export * from "./parser";
 export * from "./compiler";
 export * from "./syntax";
-export * from "./types";
+export * from "./node";
+export * from "./token";
