+++ frontend/src/metabase/querying/expressions/pratt/token.ts
@@ -1,0 +1,46 @@
+import type { NodeType } from "./node";
+
+export class Token {
+  type: NodeType;
+  text: string;
+  value?: string;
+
+  pos: number;
+  length: number;
+
+  constructor({
+    type,
+    pos,
+    length,
+    text,
+    value,
+  }: {
+    type: NodeType;
+    text: string;
+    value?: string;
+    length: number;
+    pos: number;
+  }) {
+    this.type = type;
+    this.pos = pos;
+    this.length = length;
+
+    this.text = text;
+    this.value = value;
+  }
+  get len(): number {
+    return this.length;
+  }
+  get start(): number {
+    return this.pos;
+  }
+  get end(): number {
+    return this.pos + this.length;
+  }
+  get from(): number {
+    return this.start;
+  }
+  get to(): number {
+    return this.end;
+  }
+}
